parser/requirements.txt
requests
beautifulsoup4
lxml
psycopg2-binary